# Compute proposer shuffling only once in gossip verification