--- a/common/logging/Cargo.toml
+++ b/common/logging/Cargo.toml
@@ -11,8 +11,6 @@ test_logger = [] # Print log output to stderr when running tests instead of drop
 chrono = { version = "0.4", default-features = false, features = ["clock", "std"] }
 logroller = { workspace = true }
 metrics = { workspace = true }
-once_cell = "1.17.1"
-parking_lot = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true, features = [ "time" ] }