@jimmygchen 2025-04-10
I've pushed a PoC, however I think this is still problematic because we hold the lock for the entire duration of the computation.