@jimmygchen 2025-04-29
May be worth adding this to block verification as well, as we're likely to receive the block + 128 columns within a small window under PeerDAS.