@dapplion 2025-04-29
@jimmygchen we can as well cache the proposer signature verification with a similar strategy: LRUCache + OnceCell. Have a cache `LRUCache<(BlockRoot, BlockSignature), Result>`, and use it in block, blob and data column proposer signature verification. In the supernode we will amortize the 500 ms across 128 columns.