@jimmygchen 2025-04-30
Do you see much benefit of adding this signature cache in addition to `beacon_proposer_cache`?
From the numbers above, if we hit the `beacon_proposer_cache`, the verification time is quite minimal: 0.9 - 2.5ms. Am I missing something?