@dapplion 2025-04-30
> From the numbers above, if we hit the beacon_proposer_cache, the verification time is quite minimal: 0.9 - 2.5ms. Am I missing something?

Right not necessary for now, we can revisit later if you want to hyper optimize