@jimmygchen 2025-05-01
Thanks all! 

I've found a bug in data column verification which causes proposer shuffling cache miss on every data column, due to incorrect epoch calculation only in minimal preset, as it was hard coded to use `MainnetEthSpec`.

This shouldn't happen often at all. But this PR should improve the performance in case if it happens and not do `128 + 1` state load + shuffling computation.

Agree to do it on block + blob verification as well once this is proven to be stable for some time on PeerDAS.

I'll go ahead and merge this now.